pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "Multi-agent codebase intelligence system for FDE onboarding"
requires-python = ">=3.11"
dependencies = [
    "tree-sitter>=0.23.0",
    "tree-sitter-python>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-yaml>=0.6.0",
    "sqlglot>=25.0.0",
    "networkx>=3.3",
    "pydantic>=2.7.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "gitpython>=3.1.43",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "jedi>=0.19.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "pyvis>=0.3.2",
    "requests>=2.31.0",
    "google-generativeai>=0.7.0",
    "langchain-core>=0.2.0",
    "langgraph>=0.1.0",
    "streamlit>=1.55.0",
]

[project.scripts]
cartographer = "src.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]
